Added new numeric types to the vector classes

+++ src/main/java/io/github/xasmedy/math/unit/Radians.java
@@ -1,18 +1,18 @@
 package io.github.xasmedy.math.unit;
 
 // No need for @LooselyConsistentValue, the class is 64 bit big, meaning the CPU supports atomic operations.
-public value record Radians(float value) {
+public value record Radians(double value) {
 
-    public static Radians radians(float radians) {
+    public static Radians radians(double radians) {
         return new Radians(radians);
     }
 
     /// @return A radians value converted from the provided degrees.
-    public static Radians degrees(float degrees) {
-        return radians((float) Math.toRadians(degrees));
+    public static Radians degrees(double degrees) {
+        return radians(Math.toRadians(degrees));
     }
 
-    public float toDegrees() {
-        return (float) Math.toDegrees(value);
+    public double toDegrees() {
+        return Math.toDegrees(value);
     }
 }
